+++ NEWS.md
@@ -1,6 +1,6 @@
 # Version 0.1.3
 
-*
+* Empty flag is invalid and throws an error (#3).
 
 # Version 0.1.2